Documentation des fonctions

Documentation de toute les fonctions du package Optinum

Optinum.Gradient_Conjugue_Tronque

Optinum.Lagrangien_Augmente

Optinum.Regions_De_Confiance

Optinum.Gradient_Conjugue_Tronque — Method

Minimise le problème : $min_{||s||< \delta_{k}} q_k(s) = s^{t}g + (1/2)s^{t}Hs$ pour la $k^{ème}$ itération de l'algorithme des régions de confiance

Syntaxe

sk = Gradient_Conjugue_Tronque(fk,gradfk,hessfk,option)

Entrées :

* fk               : la fonction à minimiser appliqué au point xk
* gradfk           : le gradient de la fonction f appliqué au point xk
* hessfk           : la Hessienne de la fonction f appliqué au point xk
* options
        * delta    : le rayon de la région de confiance
        * max_iter : le nombre maximal d'iterations
        * tol      : la tolerance pour la condition d'arrêt sur le gradient

Sorties:

* s : le pas s qui approche la solution du problème : ``min_{||s||< \delta_{k}} q(s)``

Exemple d'appel:

f(x)=100*(x[2]-x[1]^2)^2+(1-x[1])^2
gradf(x)=[-400*x[1]*(x[2]-x[1]^2)-2*(1-x[1]) ; 200*(x[2]-x[1]^2)]
hessf(x)=[-400*(x[2]-3*x[1]^2)+2  -400*x[1];-400*x[1]  200]
xk = [1; 0]
options = []
s = Gradient_Conjugue_Tronque(f(xk),gradf(xk),hessf(xk),options)

Optinum.Lagrangien_Augmente — Method

Resolution des problèmes de minimisation sous contraintes d'égalités

Syntaxe

Lagrangien_Augmente(algo,fonc,contrainte,gradfonc,hessfonc,grad_contrainte,
					hess_contrainte,norm_contrainte,jac_contrainte,phi,x0,option)

Entrées

algo : l'indice inqiquant l'algorithme sans contraintes à utiliser

* "newton" : pour Newton * "cauchy" : pour les régions de confiance avec le pas de Cauchy * "gct" : pour les régions de confiance avec le gradient conjugué tronqué

fonc : la fonction à minimiser

contrainte : la contrainte [x est dans le domaine des contraintes ssi $c(x)==0$]

gradfonc : le gradient de la fonction

hessfonc : la hessienne de la fonction

grad_contrainte : le gradient de la contrainte

hess_contrainte : la hessienne de la contrainte

norm_contrainte : la norme de la contrainte

jac_contrainte : la jacobienne de la contrainte

x0 : la première composante du point de départ du Lagrangien

options

* epsilon : utilisé dans les critères d'arrêt * tol : la tolérance utilisée dans les critères d'arrêt * itermax : nombre maximal d'itération dans la boucle principale * lambda0 : la deuxième composante du point de départ du Lagrangien * mu0,tho : valeurs initiales des variables de l'algorithme

Sorties

* xmin 		   : une approximation de la solution du problème avec contraintes
* fxmin 	   : ``f(x_{min}``
* flag 		   : indicateur du déroulement de l'algorithme
		0 	   : Convergence
		1 	   : nombre maximal d'itération atteint

-1 : une erreur s'est produite * niters : nombre d'itérations réalisées

Exemple d'appel

Optinum.Regions_De_Confiance — Method

Minimise une fonction en utilisant l'algorithme des régions de confiance avec - le pas de Couchy ou - le pas issu de l'algorithme du gradient conjugue tronqué

Syntaxe

xk, nb_iters, f(xk), flag = Regions_De_Confiance(algo,f,gradf,hessf,x0,option)

Entrées :

* algo        : string indicant la méthode à utiliser pour calculer le pas
                    - "gct" : pour l'algorithme du gradient conjugué tronqué
                    - "cauchy": pour le pas de Cauchy
* f           : la fonction à minimiser
* gradf       : le gradient de la fonction ``f``
* hessf       : la hessiene de la fonction à minimiser
* x0          : point de dapart
* options
  * deltaMax      : utile pour les m-à-j de la region de confiance
                    ``R_{k}=\left\{x_{k}+s ;\|s\| \leq \Delta_{k}\right\}``
  * gamma1,gamma2 : 0 < gamma1 < 1 < gamma2 pour les m-à-j de ``R_{k}``
  * eta1,eta2     : ``0 <\eta\1 < \eta_2 < 1`` pour les m-à-j de ``R_{k}``
  * delta0        : le rayon de départ de la région de confiance
  * max_iter      : le nombre maximale d'iterations
  * tol           : la tolérence pour les critères d'arrêt

Sorties:

* x_min    : une approximation de la solution du problème : ``\min_{x \in \mathbb{R}^{n}} f(x)``
* fx_min   : ``f(x_{min}``
* flag     : un entier indiquant le critère sur lequel le programme à arrêter
      0    : Convergence
      1    : stagnation du ``x``
      2    : stagnation du ``f``
      3    : nombre maximal d'itération dépassé
* nb_iters : le nombre d'iteration qu'à fait le programme

Exemple d'appel

algo="gct"
f(x)=100*(x[2]-x[1]^2)^2+(1-x[1])^2
gradf(x)=[-400*x[1]*(x[2]-x[1]^2)-2*(1-x[1]) ; 200*(x[2]-x[1]^2)]
hessf(x)=[-400*(x[2]-3*x[1]^2)+2  -400*x[1];-400*x[1]  200]
x0 = [1; 0]
option = []
xk, nb_iters, f_xk, flag = Regions_De_Confiance(algo,f,gradf,hessf,x0,option)