Documentation des fonctions

Documentation de toute les fonctions du package Optinum

Optinum.Gradient_Conjugue_Tronque

Optinum.Lagrangien_Augmente

Optinum.Regions_De_Confiance

Optinum.Gradient_Conjugue_Tronque — Method

################################################################################################### GradientConjugueTronque : minimise le problème : min q(s) avec ||s||< deltak pour l'itération k de l'algorithme des regions de confiance ###################################################################################################

#################################################################################

Entrées :

* fk               : la fonction à minimiser appliqué au point xk
* gradfk           : le gradient de la fonction f appliqué au point xk
* hessfk           : la Hessienne de la fonction f appliqué au point xk
* delta            : le rayon de la région de confiance
* max_iter         : le nombre maximal d'iterations
* tol              : la tolerance pour la condition d'arrêt sur le gradient

#################################################################################

Sorties:

* s                : le pas s qui approche la solution du problème : min q(s) avec ||s||< deltak

#################################################################################

Exemple d'appel:

* >> include("fonctions_de_tests.jl")
* >> s_min1=Gradient_Conjugue_Tronque(fct1([10;0;3]),grad_fct1([10;0;3]),hess_fct1([10;0;3]),1,[10;0;3],100,1e-15)

Optinum.Lagrangien_Augmente — Method

################################################################################# # Algorithme du lagrangien augmenté pour le calcul du minimum d'une fonction avec des contraintes d'égalités #################################################################################

################################################################################# #Entrées : * algorithmesanscontrainte : l'indice inqiquant l'algorithme sans contraintes à utiliser * "newton" : pour utiliser l'algo de Newton * "cauchy" : pour utiliser l'algo de Région de confiance avec cauchy * "gct" : pour utiliser l'algo de Région de confiance avec gradient conjugué tronqué * fonc : la fonction à minimiser * contrainte : la contrainte [x est dans le domaine des contraintes ssi c(x)==0] * gradfonc : le gradient de la fonction * hessfonc : la hessienne de la fonction * gradcontrainte : le gradient de la contrainte contrainte * hesscontrainte : la hessienne de la contrainte * nomrcontrainte : la norme de la contrainte * jaccontrainte : le jacobien de la contrainte * x0 : le point du départ * epsilon : utilisé dans les critéres d'arrêt * tol : utilisé dans les critéres d'arrêt * itermax : nombre maximal d'itération dans la boucle principale * lambda0,mu0,tho : valeurs initiales des variables de l'algorithme

#################################################################################

################################################################################# #Sorties : * xmin :Le minumum du problème avec contraintes * fxmin : l'image de xmin par la fonction objectif * niters : nombre d'itérations * flag : indicateur du déroulement de l'algorithme 0 : Convergence 1 : nombre maximal d'itération atteint -1 : une erreur s'est produite #################################################################################

Optinum.Regions_De_Confiance — Method

########################################################################### RegionDeConfiance: minimiser la fonction f en utilisant l'algorithme des régions de confiance avec - un pas de Couchy ou - un pas calculé avec l'algorithme du gradient conjugue tronque ###########################################################################

################################################################################# #Entrées : * algo : string indicant la méthode à utiliser pour calculer le pas - "gct" : pour l'algorithme du gradient conjugué tronqué - "cauchy": pour le pas de Cauchy * f : la fonction à minimiser * gradf : le gradient de la fonction f * hessf : la hessiene de la fonction à minimiser * x0 : point de dapart * deltaMax, 0 < gamma1 < 1, 1 < gamma2, 0 < eta1, eta1 < eta2 < 1 : utile pour les m-à-j de la region de confiance * delta0 : le rayon de départ * nb_itersMax : le nombre maximale d'iterations * tol : la tolérence pour les critères d'arrêt

############################################################################### #Sorties :

* x_min       : le point minimisant la fonction f
* fx_min      : la valeur minimale de la fonction f
* nb_iters    : le nombre d'iteration qu'à fait le programme
* flag        : entier indiquant le critère sur lequel le programme à arrêter
        	    0 : Convergence
        	    1 : stagnation du x
        	    2 : stagnation du f
        	    3 : nombre maximal d'itération dépassé

################################################################################